% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ced_boundaries_doc.R
\docType{data}
\name{ced_boundaries}
\alias{ced_boundaries}
\title{Australian Federal Division Electoral Boundaries}
\format{A tibble with simple features geometry}
\source{
Australian Bureau of Statistics
}
\description{
Poll and election results for first preference and two-party-preferred voting intention for the 
Australian House of Representatives, 2007 to 2019
}
\details{
a single tibble with the federal division boundaries for 2010, 2013, 2016, and 2019 federal elections.
These are called \code{ced_boundaries} to follow the ABS practice of referring to these as Commonwealth
Electoral Divisions (AEC does not use the term "Commonwealth election" but "federal election" instead).
}
