% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_boundaries_doc.R
\docType{data}
\name{div_boundaries}
\alias{div_boundaries}
\title{Australian Federal Electoral Division Boundaries}
\format{
A tibble with simple features geometry
}
\source{
Australian Bureau of Statistics
}
\description{
Federal electoral division boundaries 2010 to 2019
}
\details{
a single tibble with the federal division boundaries for 2010, 2013, 2016, and 2019 federal elections.

These maps have been simplified to 10% of their original size on the ABS website, and cropped to the bounding box
that was in use in 2010 (which excludes some far western islands that were in the 2016 and later maps),
}
