% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_boundaries_doc.R
\docType{data}
\name{div_census_2016}
\alias{div_census_2016}
\title{Australian Census by Federal Election Division}
\format{
A tibble
}
\source{
Australian Bureau of Statistics, via Hugh Parsonage's Census2016.DataPack R package, with minimal 
aggregation done by Peter Ellis for convenience.
}
\description{
Selected variables from the 2016 Census by Federal Election division
}
\details{
a single tibble with one row per electoral division and one column per variable of interest.
}
