% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{oz_party_cols}
\alias{oz_party_cols}
\title{Australian political party colours}
\format{
A named vector.
}
\source{
Party pages and Wikipedia
}
\description{
Australian political party colours
}
