% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{oz_pendulum_2019}
\alias{oz_pendulum_2019}
\title{Australian federal division margins going in to 2019 election}
\format{
A tibble with 151 rows.
}
\source{
https://www.abc.net.au/news/elections/federal-redistribution-2018/
}
\description{
Australian federal division margins going in to 2019 election
}
\details{
The margin by which each seat is held and the party against which that margin is measured.
}
