% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozpol_infographic.R
\name{ozpol_infographic}
\alias{ozpol_infographic}
\title{Draw an infographic map of Australian election results}
\usage{
ozpol_infographic(
  year,
  fontfamily = "sans",
  variable = c("lib_nat_percentage", "swing_to_govt")
)
}
\arguments{
\item{year}{Election year - should be one of 2016, 2013 or 2010}

\item{fontfamily}{Font family to use}

\item{variable}{which variable to use}
}
\description{
Draw an infographic map of Australian election results
}
\details{
This should be printed to an 8 x 6.5 inch output. Anything else might be a bit wrong... Basically,
this is a bit of a hack for a very specific single purpose. 

As a side effect, this function draws a chart that will look good on an 8 x6.5 inch output.
}
