% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{ozpolls}
\alias{ozpolls}
\title{Australian Federal Voting Intention Polls}
\format{
A tibble with 5,500+ rows and 10 variables.
}
\source{
Various polling firms, collected on Wikipedia and scraped by Peter Ellis
}
\description{
Poll and election results for first preference and two-party-preferred voting intention for the 
Australian House of Representatives, 2007 to 2019
}
