% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{ozpolls_2010}
\alias{ozpolls_2010}
\title{Australian Federal Voting Intention Polls for the 2010 Election}
\format{
A tibble with 485 rows and 9 variables.
}
\source{
Various polling firms, collected on Wikipedia and scraped by Peter Ellis
}
\description{
Poll and election results for first preference and two-party-preferred voting intention
}
\details{
Normally you don't need this as all the information in it is in the larger object \code{ozpolls}.
However, the 2010 data was unusual in having separate first preference information for Liberal and National.
This has been combined in \code{ozpolls} so the original \code{ozpolls_2010} data frame is made available if wanted.
}
