% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{ozpolls_2016}
\alias{ozpolls_2016}
\title{Australian Federal Voting Intention Polls for the 2016 Election}
\format{
A tibble with 1,518 rows and 9 variables.
}
\source{
Various polling firms, collected on Wikipedia and scraped by Peter Ellis
}
\description{
Poll and election results for first preference and two-party-preferred voting intention
}
\details{
Normally you don't need this as all the information in it is in the larger object \code{ozpolls}. 
However, the 2016 data is unusual in having information on sample size, margin of error, and method which
are not available in \code{ozpolls}.
}
