\name{pkg-package}
\alias{pkg-package}
\alias{pkg}
\docType{package}
\title{
\packageTitle{pkg}
}
\description{
\packageDescription{pkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pkg}
\packageIndices{pkg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pkg}

Maintainer: \packageMaintainer{pkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
