% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_results_doc.R
\docType{data}
\name{results_2pp_div}
\alias{results_2pp_div}
\title{Australian Federal Election Results by Division}
\format{A tibble with 750 rows and 14 variables.}
\source{
Australian Electoral Commission
}
\description{
Election results by House of Representatives Division (ie seat) 2004 to 2016
}
