% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_results_doc.R
\docType{data}
\name{results_2pp_div}
\alias{results_2pp_div}
\title{Australian Federal Election Results by Division}
\format{
A tibble with 750 rows and 14 variables.
}
\source{
Australian Electoral Commission. Results from 2004 onwards are from CSVs in the "Tally room". Earlier
results are taken from the ;-delimited text files in the bulk downloads of zip files (that used to be CD-ROMs)
}
\description{
Election results by House of Representatives Division (ie seat) 1993 to 2016
}
